requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    ./scripts/upgrade.sh all
#
alembic==1.12.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
boto3==1.29.5
botocore==1.32.5
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.8.0
docker==6.1.3
duckdb==0.9.2
executing==2.0.1
gitdb==4.0.11
gitpython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
ipython==8.17.2
jedi==0.19.1
jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
mako==1.3.0
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mypy==1.7.0
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.5
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
pgvector==0.2.3
phidata==2.0.58
pillow==10.1.0
polars==0.19.15
prompt-toolkit==3.0.41
protobuf==4.25.1
psycopg[binary]==3.1.13
psycopg-binary==3.1.13
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-settings==2.1.0
pydeck==0.8.1b0
pygments==2.17.2
pypdf==3.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
ruff==0.1.6
s3transfer==0.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.23
stack-data==0.6.3
streamlit==1.28.2
tenacity==8.2.3
tiktoken==0.5.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typer==0.9.0
types-beautifulsoup4==4.12.0.7
types-html5lib==1.1.11.15
types-pillow==10.1.0.2
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.12
websocket-client==1.6.4
zipp==3.17.0
